requirements.txt
asgiref==3.7.2
build==1.1.1
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0
django-easyjwt==0.4.1
djangorestframework==3.14.0
filelock==3.13.1
identify==2.5.35
idna==3.6
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pre-commit==3.6.2
pyproject_hooks==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.3
sqlparse==0.4.4
tomli==2.0.1
types-requests==2.31.0.20240311
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.1
